moved tabs object into Tabs component and use Outlet to display Tab
