+++ src/App.tsx
@@ -4,18 +4,13 @@ import './App.scss';
 import {
   Navigate,
   NavLink,
+  Outlet,
   Route,
   Routes,
   useLocation,
 } from 'react-router-dom';
 import { Tabs } from './Components/Tabs';
 import classNames from 'classnames';
-
-const tabs = [
-  { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
-  { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
-  { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
-];
 
 export const App = () => {
   const { pathname } = useLocation();
@@ -59,8 +54,8 @@ export const App = () => {
           {pathname === '/home' && <Navigate to="/" />}
           <Routes>
             <Route path="/" element={<h1 className="title">Home page</h1>} />
-            <Route path="/tabs" element={<Tabs tabs={tabs} />}>
-              <Route path=":tabId" element={<Tabs tabs={tabs} />} />
+            <Route path="/tabs" element={<Tabs />}>
+              <Route path=":tabId" element={<Outlet />} />
             </Route>
             <Route
               path="*"
