+++ src/Components/Tabs/TabContent/index.ts
@@ -1,0 +1,1 @@
+export * from './TabContent';
