+++ src/Components/Tabs/Tabs.tsx
@@ -1,14 +1,12 @@
-import { Link, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
+import { TabContent } from './TabContent';
+const tabs = [
+  { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
+  { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
+  { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
+];
 
-type Props = {
-  tabs: {
-    id: string;
-    title: string;
-    content: string;
-  }[];
-};
-
-export const Tabs = ({ tabs }: Props) => {
+export const Tabs = () => {
   const { tabId } = useParams<{ tabId?: string }>();
   const activeTab = tabs.find(tab => tab.id === tabId);
   // move routing between tabs inside this component
@@ -18,19 +16,7 @@ export const Tabs = ({ tabs }: Props) => {
       <h1 className="title">Tabs page</h1>
       <div className="tabs is-boxed">
         <ul>
-          {tabs.map(tab => {
-            return (
-              <li
-                key={tab.id}
-                className={tabId === tab.id ? 'is-active' : ''}
-                data-cy="Tab"
-              >
-                <Link to={tab.id} data-cy="TabLink">
-                  {tab.title}
-                </Link>
-              </li>
-            );
-          })}
+          <TabContent tabs={tabs} tabId={tabId} />
         </ul>
       </div>
       <div className="block" data-cy="TabContent">
